feat. service 및 controller 일부 수정
