+++ src/main/java/com/concert_reservation/api/application/facade/BookingFacade.java
@@ -9,14 +9,10 @@ import com.concert_reservation.api.application.dto.request.SeatRequest;
 import com.concert_reservation.api.application.dto.response.AvailableDatesResponse;
 import com.concert_reservation.api.application.dto.response.BookingResponse;
 import com.concert_reservation.api.application.dto.response.SeatResponse;
-import com.concert_reservation.api.business.model.dto.command.BookingCommand;
 import com.concert_reservation.api.business.model.dto.command.SeatCommand;
 import com.concert_reservation.api.business.model.dto.info.BookingInfo;
 import com.concert_reservation.api.business.model.dto.info.SeatInfo;
-import com.concert_reservation.api.business.model.dto.info.UserInfo;
 import com.concert_reservation.api.business.service.BookingService;
-import com.concert_reservation.api.business.service.impl.BookingServiceImpl;
-import com.concert_reservation.api.business.service.impl.UserServiceImpl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
