+++ src/main/java/com/concert_reservation/api/application/facade/ConcertFacade.java
@@ -3,8 +3,9 @@ package com.concert_reservation.api.application.facade;
 // import com.concert_reservation.api.business.service.impl.ConcertServiceImpl;
 import com.concert_reservation.api.business.service.ConcertService;
 import com.concert_reservation.api.business.service.UserService;
-import com.concert_reservation.api.business.service.impl.UserServiceImpl;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -12,7 +13,6 @@ import com.concert_reservation.api.application.dto.request.ConcertRequest;
 import com.concert_reservation.api.application.dto.response.ConcertResponse;
 import com.concert_reservation.api.business.model.dto.command.ConcertCommand;
 import com.concert_reservation.api.business.model.dto.info.ConcertInfo;
-import com.concert_reservation.api.business.service.impl.ConcertServiceImpl;
 
 @Component
 @RequiredArgsConstructor
@@ -50,5 +50,9 @@ public class ConcertFacade {
 		concertService.deleteConcert(concertId);
 	}
 
+	public List<ConcertResponse> getConcerts() {
+		List<ConcertInfo> concertInfos = concertService.getConcerts();
+		return concertInfos.stream().map(ConcertResponse::from).collect(Collectors.toList());
+	}
 
 }
