--- src/main/java/com/concert_reservation/api/application/facade/PointFacade.java
@@ -1,20 +1,0 @@
-package com.concert_reservation.api.application.facade;
-
-import org.springframework.stereotype.Component;
-
-import com.concert_reservation.api.application.dto.request.PaymentRequest;
-import com.concert_reservation.api.application.dto.request.PointRequest;
-import com.concert_reservation.api.application.dto.response.PaymentResponse;
-import com.concert_reservation.api.application.dto.response.PointResponse;
-import com.concert_reservation.api.business.model.dto.command.PointCommand;
-import com.concert_reservation.api.business.model.dto.info.PointInfo;
-import com.concert_reservation.api.business.service.PointService;
-
-import lombok.RequiredArgsConstructor;
-
-@Component
-@RequiredArgsConstructor
-public class PointFacade {
-
-
-}
