+++ src/main/java/com/concert_reservation/api/business/service/TokenServiceImpl.java
@@ -1,4 +1,4 @@
-package com.concert_reservation.api.business.service.impl;
+package com.concert_reservation.api.business.service;
 
 import java.time.LocalDateTime;
 import java.util.List;
