+++ src/main/java/com/concert_reservation/api/infrastructure/persistance/orm/WaitingCountJpaRepository.java
@@ -1,11 +1,8 @@
 package com.concert_reservation.api.infrastructure.persistance.orm;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.transaction.annotation.Transactional;
-
 import com.concert_reservation.api.business.model.entity.WaitingCount;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 public interface WaitingCountJpaRepository extends JpaRepository<WaitingCount, Long> {
